updated release to 0.6.2, DBI updates to handle databases that require explicit BEGINs for transaction management, commented out unused gz_header as it causes compile problems with older zlib headers
git-svn-id: svn+ssh://svn.code.sf.net/p/qore/code/qore/trunk@764 1b0b35fe-f740-4c1b-b404-a338b32e513b

+++ lib/QoreType.cc
@@ -429,7 +429,7 @@ void QoreTypeManager::init()
    Zero        = new QoreNode((int64)0);
    ZeroFloat   = new QoreNode(0.0);
    NullString  = new QoreNode("");
-   ZeroDate    = new QoreNode(new DateTime(0ll));
+   ZeroDate    = new QoreNode(new DateTime((int64)0));
    
    emptyList   = new QoreNode(new List());
    emptyHash   = new QoreNode(new Hash());
